third_party/blink/web_tests/platform/mac/external/wpt/permissions-policy/payment-extension-allowed-by-permissions-policy-attribute.https.sub-expected.txt

This is a testharness.js-based test.
PASS permissions policy "payment" is not supported in cross-origin iframe without allow="payment" attribute
FAIL permissions policy "payment" can be enabled in cross-origin iframe using allow="payment" attribute assert_equals: expected "OK" but got "NotSupportedError#The user agent does not support public key credentials."
Harness: the test ran to completion.
